Move files into app folder
