+++ app/config.py
@@ -1,0 +1,10 @@
+## Config options. Enter your Discord bot and Plex information. ##
+
+## Discord Bot Token ##
+TOKEN = "Enter your Discord bot token here"
+
+## Plex Token ##
+PLEX_TOKEN = "Enter your Plex token here"
+
+## Plex base URL ##
+BASEURL = "ENTER YOUR PLEX URL HERE"
